Fix compilation warnings (#1822)
Resolve multiple compilation warnings:
- unused argument
- type conversion

+++ src/apps/relay/dtls_listener.c
@@ -155,7 +155,7 @@ static size_t print_packet_txt2pcap(uint64_t now, uint8_t *payload, size_t paylo
   index += snprintf((char *)(txt2pcap + index), txt2pcap_length - index, " 0000");
 
   for (size_t i = 0; i < payload_length; i++) {
-    int n = snprintf((char *)(txt2pcap + index), txt2pcap_length - index, " %02x", payload[i]);
+    size_t n = snprintf((char *)(txt2pcap + index), txt2pcap_length - index, " %02x", payload[i]);
     if (n < 0 || n >= txt2pcap_length - index) {
       break;
     }
